backend/requirements.txt
Flask==3.1.2
gunicorn==23.0.0
flask-cors==6.0.1
requests==2.32.5
pandas==2.3.3
prophet==1.2.1
numpy==1.23.5
joblib==1.5.2
python-dotenv==1.2.1
psycopg2-binary==2.9.7
google-cloud-storage==2.11.0

cloud-sql-python-connector[pg8000]==1.3.0
pg8000==1.29.8

tensorflow==2.12.0
SQLAlchemy>=1.4,<2

